refactor (Lua include package path to scriptdir)
