+++ shttps/LuaServer.cpp
@@ -2189,6 +2189,20 @@ namespace shttps {
     //=========================================================================
 
 
+    void LuaServer::setLuaPath(const string &path) {
+        lua_getglobal(L, "package" );
+        lua_getfield(L, -1, "path" ); // get field "path" from table at top of stack (-1)
+        std::string cur_path = lua_tostring( L, -1 ); // grab path string from top of stack
+        cur_path.append( ";" ); // do your path magic here
+        cur_path.append( path );
+        lua_pop( L, 1 ); // get rid of the string on the stack we just pushed on line 5
+        lua_pushstring( L, cur_path.c_str() ); // push the new one
+        lua_setfield( L, -2, "path" ); // set the field "path" in table at -2 with value at top of stack
+        lua_pop( L, 1 ); // get rid of package table from top of stack
+        return; // all done!
+    }
+
+
     /*!
      * This function registers all variables and functions in the server table
      */
@@ -2658,9 +2672,15 @@ namespace shttps {
 
     int LuaServer::executeChunk(const string &luastr) {
         if (luaL_dostring(L, luastr.c_str()) != LUA_OK) {
-            const char *errormsg = lua_tostring(L, 1);
-            lua_pop(L, 1);
-            throw Error(__file__, __LINE__, string("LuaServer::executeChunk failed: ") + errormsg);
+            const char *errormsg = NULL;
+            if (lua_gettop(L) > 0) {
+                errormsg = lua_tostring(L, 1);
+                lua_pop(L, 1);
+                throw Error(__file__, __LINE__, string("LuaServer::executeChunk failed: ") + errormsg);
+            }
+            else {
+                throw Error(__file__, __LINE__, "LuaServer::executeChunk failed!");
+            }
         }
         int top = lua_gettop(L);
         if (top == 1) {
