+++ shttps/Server.cpp
@@ -1087,6 +1087,7 @@ namespace shttps {
             // Setting up the Lua server
             //
             LuaServer luaserver(conn, _initscript, true);
+            luaserver.setLuaPath(_scriptdir + "/?.lua"); // add the script dir to the standard search path fpr lua packages
             //luaserver.createGlobals(conn);
             for (auto &global_func : lua_globals) {
                 global_func.func(luaserver.lua(), conn, global_func.func_dataptr);
